Adding a hook for computation of the soil source
This adds a post tendency hook that allows for computing quantities that
might be easiest to calculate after the tendency computation.

The current use case is to compute a soil source term that depends on
`|∇κ∇T|`. The problem is that we need to use `∇κ∇T`, and its full
computation isn't until *after* the full tendency is evaluated. Also,
the quantity itself is never stored (though it is already calculated).
By allowing a new hook after the tendency we can piggyback on the
previous computation (at the hazard of addition a possibly dangerous
hole) in the tendency evaluation.

The alternate approach would be to add an auxiliary balance law, compute
this in the update auxiliary and then compute the actual tendency. But
doing this would add a lot more looping through the data and computing
the same quantity multiple times.

Co-authored-by: Katherine Deck <kdeck@caltech.edu>